restructure: move local_storage functions
