+++ microdata_validator/__init__.py
@@ -12,7 +12,7 @@ from microdata_validator.steps import (
     dataset_reader, dataset_validator
 )
 from microdata_validator.schema import validate_with_schema
-
+from microdata_validator.repository import local_storage
 
 logger = logging.getLogger()
 
@@ -115,7 +115,7 @@ def validate_metadata(metadata_file_path: str,
     try:
         metadata_file_path = Path(metadata_file_path)
         if metadata_ref_directory is None:
-            metadata_dict = utils.load_json(Path(metadata_file_path))
+            metadata_dict = local_storage.load_json(Path(metadata_file_path))
         else:
             metadata_ref_directory = Path(metadata_ref_directory)
             metadata_dict = utils.inline_metadata_references(
@@ -155,7 +155,7 @@ def inline_metadata(metadata_file_path: str, metadata_ref_directory: str,
     )
     validate_with_schema(metadata_dict)
 
-    utils.write_json(output_file_path, metadata_dict)
+    local_storage.write_json(output_file_path, metadata_dict)
     return output_file_path
