+++ tests/test_inline_metadata.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import pytest
 
 from microdata_validator import inline_metadata
+from microdata_validator.repository import local_storage
 from microdata_validator import utils
 from microdata_validator.utils import ParseMetadataError
 
@@ -35,8 +36,8 @@ def test_inline_full_dataset():
         REF_DIR,
         output_file_path=OUTPUT_FILE_PATH
     )
-    actual_inlined = utils.load_json(result_file_path)
-    assert actual_inlined == utils.load_json(
+    actual_inlined = local_storage.load_json(result_file_path)
+    assert actual_inlined == local_storage.load_json(
         Path(SIVSTAND_INLINE_FILE_PATH)
     )
 
@@ -55,8 +56,8 @@ def test_inline_full_dataset_default_output_path():
         REF_DIR
     )
     assert str(result_file_path) == KJOENN_DEFAULT_OUTPUT_PATH
-    actual_inlined = utils.load_json(result_file_path)
-    assert actual_inlined == utils.load_json(
+    actual_inlined = local_storage.load_json(result_file_path)
+    assert actual_inlined == local_storage.load_json(
         Path(KJOENN_INLINED_FILE_PATH)
     )
